change to prompts and add hook for click outside to close
